Shuffle logic to facades / services
