+++ app/controllers/api/v1/forecast_controller.rb
@@ -1,14 +1,7 @@
 class Api::V1::ForecastController < ApplicationController
 
   def show
-    forecast = weather.forecast(params[:location])
-    
-    render json: forecast
+    render json: WeatherFacade.forecast(params[:location])
   end
 
-  private
-
-  def weather
-    @_weather = WeatherService.new
-  end
 end
