+++ app/facades/base_facade.rb
@@ -1,0 +1,4 @@
+class BaseFacade
+
+
+end
