+++ app/facades/weather_facade.rb
@@ -1,0 +1,12 @@
+class WeatherFacade < BaseFacade
+
+  def self.forecast(location)
+    weather.forecast(location)
+  end
+  private
+
+  def self.weather
+    @_weather = WeatherService.new
+
+  end
+end
