+++ spec/requests/api/v1/users/favorite_locations_spec.rb
@@ -19,7 +19,7 @@ describe Api::V1::FavoritesController do
 
     expect(response).to be_successful
     expect(user.favorites.last.location).to eq("Denver, CO")
-    # binding.pry
+    # #binding.pry
   end
 
   it 'can get favorite locations' do
